Creating a helper file for feds utils function
