+++ test/features/georoutingv2/georoutingv2.test.js
@@ -3,9 +3,9 @@ import { expect } from '@esm-bundle/chai';
 import { setViewport } from '@web/test-runner-commands';
 
 const { default: init, getCookie } = await import('../../../libs/features/georoutingv2/georoutingv2.js');
-
 let { getMetadata } = await import('../../../libs/utils/utils.js');
-const { createTag, loadStyle, loadBlock, setConfig, getFederatedContentRoot } = await import('../../../libs/utils/utils.js');
+const { getFederatedContentRoot } = await import('../../../libs/utils/fedsHelpers.js');
+const { createTag, loadStyle, loadBlock, setConfig } = await import('../../../libs/utils/utils.js');
 
 const mockConfig = {
   locales: {
